practice

+++ class1/arithmatic.java
@@ -2,7 +2,7 @@ package class1;
 
 public class arithmatic {
     public static void main(String[] args) {
-        int x=5, y=7;
+        double x=5, y=7;
         System.out.println(x+y);
         System.out.println(x-y);
         System.out.println(x*y);
